+++ core/impl/src/main/java/com/paranid5/star_wars_travel/impl/presentation/SocInfo.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 @Immutable
 @Parcelize
 data class SocInfo(
-    val population: Int,
+    val population: Long,
     val nativeSpecies: Array<String> = emptyArray(),
     val otherSpecies: Array<String> = emptyArray(),
     val primaryLanguages: Array<String> = emptyArray(),
